Refactored profile components to use shared S3 components
Replaced custom S3 logic in all profile-related components with shared upload/display components.
