+++ src/components/shared/index.ts
@@ -1,0 +1,5 @@
+// Shared components for file uploads and profile management
+export { ProfilePhotoDisplay } from './profile-photo-display';
+export { ProfilePhotoUpload } from './profile-photo-upload';
+export { DocumentAttachmentGrid } from './document-attachment-grid';
+export { DocumentUploadSection } from './document-upload-section';
